Moving shared types to utils
