+++ src/utils/hrToMs.ts
@@ -1,4 +1,4 @@
-import { HrTime } from '../types';
+import { HrTime } from './types';
 
 export default ([seconds, nanoseconds]: HrTime): number =>
 	seconds + nanoseconds / 1e9;
